% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-isochrone.R
\name{approx_grid}
\alias{approx_grid}
\title{utility functions used in make_distance_isochrone.R and make_time_isochrone.R}
\usage{
approx_grid(lat, lng, km)
}
\description{
utility functions used in make_distance_isochrone.R and make_time_isochrone.R
}
