% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoCode.R
\name{geocode}
\alias{geocode}
\title{geocode a location using the google maps geocode API.}
\usage{
geocode(address, api_key = Sys.getenv("google_api_key"),
  verbose = FALSE, return_all = TRUE)
}
\arguments{
\item{address}{character string with address to be geocoded. Be as specific as possible to avoid ambiguity.}

\item{api_key}{character string containing api key for the google maps geocode api. You can generate one for free in your google cloud console.}

\item{verbose}{logical value. If TRUE, echos the input address}

\item{return_all}{logical value. Some locations may return more than one geocoded result. If you only want one row returned set return_all = FALSE.}
}
\value{
a data.frame containing the coordinates, location type and approximate address.
}
\description{
The function geocodes a single address given as a character string. Be as specific as possible, including the country or postcode to avoid ambiguity.
To avoid loss of accuracy when geocoding a lat/lng string, the geoCode function does not geocode lat/lng strings but instead returns the same string as a dataframe.
}
\examples{
geocode('London Bridge, London, UK', return_all = FALSE)

#  lat        lng             type                           address
#  51.50788 -0.0877321 GEOMETRIC_CENTER London Bridge, London SE1 9RA, UK

}
