% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_employment_within.R
\name{get_employment_within}
\alias{get_employment_within}
\title{Get Employment Within Boundary}
\usage{
get_employment_within(boundary, year, type = "employment", split = F,
  industry = "all", api_key = Sys.getenv("nomis_api_key"))
}
\arguments{
\item{boundary}{an sf polygon in England or Wales.}

\item{year}{the year for which to return employment data}

\item{type}{'employment' or 'employees'}

\item{split}{If split is set to TRUE and type is set to 'employees', the employment will be separated into part time and full time workers}

\item{industry}{'all' for no industry breakdown. 'broad' for broad sections, 'sections' for sections, or '2digit' for two digit SIC codes.}

\item{api_key}{can be specified manually here, or set using 'set_nomis_api' function.}
}
\value{
a tibble containing employment data
}
\description{
Fetches the employment within a given boundary. The boundary is given as an sf file and intersected with LSOAs.
The employment data is fetched from NOMIS Business Register and Employment Survey (BRES) data.
}
\examples{
emp <- get_employment_within(shp, 2018, industry = 'all')


}
