% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population_within.R
\name{get_population_within}
\alias{get_population_within}
\title{Get Population Within Boundary}
\usage{
get_population_within(
  boundary,
  year = "latest",
  age = "all",
  api_key = Sys.getenv("nomis_api_key")
)
}
\arguments{
\item{boundary}{an sf polygon in England or Wales.}

\item{year}{the year for which to return employment data}

\item{age}{'all' for no age breakdown. 'working' returns young, working age, and older groupings. 'five_year' for five year groupings or 'sya' for single year of age.}

\item{api_key}{can be specified manually here, or set using 'set_nomis_api' function.}
}
\value{
a tibble containing employment data
}
\description{
Fetches the population within a given boundary. The boundary is given as an sf file and intersected with LSOAs.
The population data is fetched from NOMIS population estimates for small areas.
}
\examples{
iso <- make_isochrone(site = 'SE1 9SG', time = 30, method = 'mapbox', mode = 'driving')
pop <- get_population_within(iso, 'latest', age = 'all')
# total employment within isochrone:
pop \%>\%
 dplyr::summarise(total_pop = sum(population_within))

}
