% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_isodistance.R
\name{make_isodistance}
\alias{make_isodistance}
\title{Make Isodistance}
\usage{
make_isodistance(site, direction = "out", distance, pts = 10,
  mode = "driving", api_key = Sys.getenv("google_api_key"))
}
\arguments{
\item{site}{either a data.frame with the columns 'lat' and 'lng' or a character string to be geocoded.}

\item{direction}{either 'in' or 'out'. In generates an isochrone from multiple origins to the site. Out goes from the site to multiple destinations.}

\item{distance}{a numeric value in meters - the distance to the isochrone extents.}

\item{pts}{the number of points to include in estimating our isochrone. More points will be more accurate but will eat up more queries from the google maps API allownce. It is recommended to test on a low number ~10 or less to ensure all of the parameters are correct before attempting to increase the number of points.}

\item{mode}{a character string for the mode of travel. Possible values are 'driving', 'cycling', 'transit', or 'walking'}

\item{api_key}{the google maps API key. This can be generated from the google cloud console.}
}
\value{
a list containing the data.frame, leaflet map, and sf object generated.
}
\description{
Generates an isodistance polygon for travel times to/from a specified location.
}
\examples{
distance_isochrone <- make_distance_isochrone(site = 'London Bridge Station, London, UK', 'out', distance = 1609, mode = 'walking', api_key = api_key)


}
