% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_google_api.R
\name{set_google_api}
\alias{set_google_api}
\title{Set API Key}
\usage{
set_google_api(key)
}
\arguments{
\item{key}{the api key for google cloud console. It must be enabled for the google distance matrix API and geocoding API.}
}
\description{
Helper function to set the google API key as an environment variable
}
\examples{
\dontrun{
set_google_api('<api key here>')
# Check our assignment worked
Sys.getenv('google_api_key')

#[1] "<api key here>"
}
}
