% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mapbox_api.R
\name{set_mapbox_api}
\alias{set_mapbox_api}
\title{Set API Key}
\usage{
set_mapbox_api(key)
}
\arguments{
\item{key}{a mapbox API key}
}
\description{
Helper function to set the mapbox API key as an environment variable
}
\examples{
\dontrun{
set_mapbox_api('<api key here>')
# Check our assignment worked
Sys.getenv('mapbox_api_key')

#[1] "<api key here>"
}
}
