% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_nomis_api.R
\name{set_nomis_api}
\alias{set_nomis_api}
\title{Set API Key}
\usage{
set_nomis_api(key)
}
\arguments{
\item{key}{the NOMIS API Key (unique ID)}
}
\description{
Helper function to set the NOMIS API key as an environment variable
}
\examples{
set_nomis_api('<api key here>')
# Check our assignment worked
Sys.getenv('nomis_api_key')

#[1] "<api key here>"

}
