% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-geocode.R
\name{url}
\alias{url}
\title{Utility functions used in geoCode.R}
\usage{
url(address, return.call = "json", sensor = "false",
  api_key = "AIzaSyB5N4Uij7F5XchUWJ-xvCqLBVfnmWNj9f0")
}
\description{
Utility functions used in geoCode.R
}
