% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-isodistance.R
\name{validate_distance_args}
\alias{validate_distance_args}
\title{utility functions used in make_distance_isochrone.R and make_time_isochrone.R}
\usage{
validate_distance_args(google_api_key, direction, site, distance, detail,
  mode, departing)
}
\description{
utility functions used in make_distance_isochrone.R and make_time_isochrone.R
}
