% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_isodistance.R
\name{make_isodistance}
\alias{make_isodistance}
\title{Make Isodistance}
\usage{
make_isodistance(site, distance, direction = c("out", "in"),
  detail = "medium", mode = c("driving", "walking", "cycling",
  "transit"), departing = FALSE,
  google_api_key = Sys.getenv("google_api_key"))
}
\arguments{
\item{site}{either a data.frame with the columns 'lat' and 'lng' or a character string to be geocoded.}

\item{distance}{a numeric value in meters - the distance to the isochrone extents.}

\item{direction}{either 'in' or 'out'. In generates an isochrone from multiple origins to the site. Out goes from the site to multiple destinations.}

\item{detail}{'low', 'medium' or 'high' level of detail in the isochrone. High will produce the most granular detail but will use more API credits.}

\item{mode}{a character string for the mode of travel. Possible values are 'driving', 'cycling', 'transit', or 'walking'}

\item{departing}{optional parameter for getting distance in traffic. Google maps may route differently to avoid heavy traffic, changing the distance at peak times. If set, takes the format "YYYY-MM-DD HH:MM:SS"}

\item{google_api_key}{the google maps API key. This can be generated from the google cloud console and set with set_google_api}
}
\value{
an sf polygon
}
\description{
Generates an isodistance polygon for network distances to/from a specified location.
}
\examples{
walk_radius <- make_isodistance('EC2R 8AH', distance = 2000, direction = 'out', mode = 'walking')


}
